Integration tests now working again locally and in the CI pipeline.
Tried to make it as simple as possible. Will need improvements,
probably. And the timing of the postgres container being ready before
the tests start running ma become a problem... In other machines maybe,
or under a different load, maybe...

We may need to go back to using wait-for-it.sh or a similar solution.