Using only not-deleted operations and transfers when calculating the balances. Maybe?
I almost do not remember how any of this works! But that's just as
important as everything else in this experience. If I can change and
update bits and pieces having a reasonable-ish idea of what I'm doing,
safely, without crashing everything, that's a very good sign.

Let's see what the next commit will say! =P